requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiosqlite==0.19.0
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
xgboost==2.0.3

